+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 2.2.6
 - Add Spanish (Chile) Translations (es_cl) (Credits: Ganbare-Lucifer)
-- Generalize Particle Registration
+- Generalize Particle, Block and Item Color Registrations
 - Finish and Tweak German Translations (de_de) (Credits: ieguana)
 - Fix StemBlock/AttachedStemBlock/PitcherCropBlock not being placeable on BWG Farmland (Fabric Only)